% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[golem]{with_golem_options}}.}
}
\description{
Run the Shiny Application
}
